code/kfserving/kfserving/requirements.txt
kfserving==0.3.0
numpy
tensorflow==2.0.0-alpha0
Pillow
requests